Rename data to model
