+++ functions/src/image/fetch-image.test.ts
@@ -1,6 +1,6 @@
 import * as admin from 'firebase-admin'
 import * as test from 'firebase-functions-test'
-import { Collections } from '../data/collections'
+import { Collections } from '../model/collections'
 import * as functions from '../index'
 
 describe('FetchImage', () => {
